@@ -25,12 +25,12 @@
     "electron": "npx cap open @capacitor-community/electron"
   },
   "dependencies": {
-    "@angular/common": "~13.0.3",
-    "@angular/core": "~13.0.3",
-    "@angular/forms": "~13.0.3",
-    "@angular/platform-browser": "~13.0.3",
-    "@angular/platform-browser-dynamic": "~13.0.3",
-    "@angular/router": "~13.0.3",
+    "@angular/common": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/forms": "^14.2.12",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
     "@capacitor/app": "1.0.6",
     "@capacitor/camera": "^1.2.1",
     "@capacitor/core": "3.3.2",
@@ -47,11 +47,11 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.0.4",
-    "@angular/cli": "~13.0.4",
-    "@angular/compiler": "~13.0.3",
-    "@angular/compiler-cli": "~13.0.3",
-    "@angular/language-service": "~13.0.3",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler": "^14.2.12",
+    "@angular/compiler-cli": "^14.2.12",
+    "@angular/language-service": "^14.2.12",
     "@capacitor-community/electron": "^4.0.3",
     "@capacitor/android": "^3.3.2",
     "@capacitor/cli": "^3.3.2",
@@ -66,6 +66,6 @@
     "karma-coverage": "^2.1.0",
     "karma-jasmine": "^4.0.1",
     "karma-jasmine-html-reporter": "^1.7.0",
-    "typescript": "~4.3.5"
+    "typescript": "~4.8.4"
   }
 }
